requirements.txt
requests==2.32.3
fastapi[standard]==0.115.0
uvicorn==0.30.6
openai==1.50.2
python-dotenv==1.0.1
